8388791: Refactor JavaClasses Throwable and StackTrace classes
