+++ src/hotspot/share/classfile/javaClassesImpl.hpp
@@ -27,6 +27,7 @@
 
 #include "cds/serializeClosure.hpp"
 #include "classfile/javaClasses.hpp"
+#include "classfile/javaStackTraceClasses.hpp"
 #include "runtime/continuationJavaClasses.hpp"
 #include "utilities/macros.hpp"
