+++ src/hotspot/share/jfr/dcmd/jfrDcmds.cpp
@@ -23,6 +23,7 @@
  */
 
 #include "classfile/javaClasses.inline.hpp"
+#include "classfile/javaStackTraceClasses.hpp"
 #include "classfile/vmSymbols.hpp"
 #include "jfr/dcmd/jfrDcmds.hpp"
 #include "jfr/jfr.hpp"
