+++ src/hotspot/share/jfr/recorder/jfrRecorder.cpp
@@ -23,7 +23,7 @@
  */
 
 #include "cds/cdsConfig.hpp"
-#include "classfile/javaClasses.hpp"
+#include "classfile/javaStackTraceClasses.hpp"
 #include "jfr/dcmd/jfrDcmds.hpp"
 #include "jfr/instrumentation/jfrJvmtiAgent.hpp"
 #include "jfr/jni/jfrJavaSupport.hpp"
