+++ src/hotspot/share/jfr/recorder/service/jfrOptionSet.cpp
@@ -24,6 +24,7 @@
 
 #include "cds/cdsConfig.hpp"
 #include "classfile/javaClasses.hpp"
+#include "classfile/javaStackTraceClasses.hpp"
 #include "jfr/dcmd/jfrDcmds.hpp"
 #include "jfr/periodic/jfrRedactedEvents.hpp"
 #include "jfr/recorder/service/jfrMemorySizer.hpp"
