+++ src/hotspot/share/runtime/sharedRuntime.cpp
@@ -28,6 +28,7 @@
 #include "classfile/classLoader.hpp"
 #include "classfile/compactHashtable.hpp"
 #include "classfile/javaClasses.inline.hpp"
+#include "classfile/javaStackTraceClasses.hpp"
 #include "classfile/stringTable.hpp"
 #include "classfile/vmClasses.hpp"
 #include "classfile/vmSymbols.hpp"
