+++ src/hotspot/share/utilities/exceptions.cpp
@@ -23,6 +23,7 @@
  */
 
 #include "classfile/javaClasses.hpp"
+#include "classfile/javaStackTraceClasses.hpp"
 #include "classfile/systemDictionary.hpp"
 #include "classfile/vmClasses.hpp"
 #include "classfile/vmSymbols.hpp"
